RSVProgramme Class

A class to represent an RSV (Respiratory Syncytial Virus) Programme.

Slots

econ_name

A character string representing the economic name.

prog_name

A character string representing the programme name.

model

An object of class refClass representing the model.

uk_data

A list containing UK data.

model_par

A list of model parameters.

econ_par

A list of economic parameters.

model_calendar

A list representing the model calendar.

dose_calendar

A list representing the dose calendar.

immune_profile

A list representing the immune profile.

econ_df

A data.frame containing economic data.

risks_df

A data.frame containing risk data.

risks_vhr_df

A data.frame containing very high risk data.

outcomes_vec

A vector of outcomes.

post

A data.frame containing posterior data.

seeds

A vector of seeds.

S

A numeric value.

raw_inci

A list of raw incidence data.

wane_func_string

A character string representing the waning function.

wane_func_string_vhr

A character string representing the waning function for very high risk.

outcomes

A list of outcomes.

samples_outcomes

A list of sample outcomes.

samples_outcomes_VHR

A list of sample outcomes for very high risk.

efficacies

A vector of efficacies.

cov_mat

A numeric matrix representing the covariance matrix.

full_output

A logical value indicating whether to produce full output.